Revert "universal8890: copy libGLES_mali.so to vulkan.exynos5.so"
 * apparently symlinks works again?

This reverts commit 6fecc7c5c8d889496557d520c9a2c7056e22ae58.

[fcuzzocrea] also move the vulkan hal outside hw folder to workaround

E vndksupport: Could not load vulkan.exynos5.so from default namespace: dlopen failed: library "vulkan.exynos5.so" not found.

since commit
https://github.com/LineageOS/android_frameworks_native/commit/d10af92a7b83d252038944cad6014a00bfa7ca25
the system tries to dlopen vulkan hal from the default namespace, which
is not allowed to search for libs in /vendor/{LIB}/hw. Extending the
default namespace to access lib(64)/hw is more hacky than just putting
a symlink in lib(64), so go with that route.

Change-Id: I7bee55e415bb30f2fd72bb88480302ef48f72818